fs_fill: Fix test when running on a 256 CPU+ machine
The fs_fill test runs a fill test on all the supported filesystems. One
of them being vfat. This filesystem is configured dynamically or through
flags/arguments for its file allocation table type (12/16/32).

The size of the test device (which is a loop-mounted fs) is 300MB. When
not instructed, mkfs will "automatically select between 12, 16 and 32
bit, whatever fits better for the filesystem size"[1]. In the case of
a 300Mb that would end up as FAT16.

This is linked with another configuration that is the actual impact on
this issue: the maximum number of directories in the root of the
filesystem. FAT12 and FAT16 use a special system region: "Root Directory
Region". And by default (there is also an argument to configure this at
mkfs-time) this ends up being 256 when no custom arguments are provided.

On the other hand, the test sets up the filesystem with a
"tst_ncpus_conf + 2" number of directories in the test filesystem's root
which can break the limit explained above on systems with a number of
CPUs that would amount above the limit.

This change addresses this situation by using a subdirectory in the test
filesystem which is not subject to this limit.

Signed-off-by: Andrei Gherzan <andrei.gherzan@canonical.com>